proseed

+++ news/src/main/java/com/igr76/news/service/FeedService.java
@@ -4,5 +4,5 @@ import com.igr76.news.dto.FeedDto;
 import com.igr76.news.dto.UpdateFeedDto;
 
 public interface FeedService {
-    FeedDto path(Long id, UpdateFeedDto updateFeedDto);
+    FeedDto path(int id, UpdateFeedDto updateFeedDto);
 }
